Revert "Integrate index op (#83)" (#88)
This reverts commit e866d3e47875accbb074d941469170869bd60683.

+++ hyracks-fullstack/hyracks/hyracks-storage-am-btree/src/main/java/org/apache/hyracks/storage/am/vector/api/IVectorClusteringMetadataFrame.java
@@ -54,7 +54,7 @@ public interface IVectorClusteringMetadataFrame extends IVectorClusteringFrame {
      * @return the maximum distance
      * @throws HyracksDataException if an error occurs
      */
-    double getMaxDistance(int tupleIndex) throws HyracksDataException;
+    float getMaxDistance(int tupleIndex) throws HyracksDataException;
 
     /**
      * Gets the data page pointer for a metadata entry.
@@ -79,7 +79,7 @@ public interface IVectorClusteringMetadataFrame extends IVectorClusteringFrame {
      * @return ITupleReference representing the metadata tuple
      * @throws HyracksDataException if tuple creation fails
      */
-    ITupleReference createMetadataTuple(double maxDistance, int dataPageId) throws HyracksDataException;
+    ITupleReference createMetadataTuple(float maxDistance, int dataPageId) throws HyracksDataException;
 
     /**
      * Finds the index where a new tuple should be inserted.
